requirements.txt
flask==3.0.3
gunicorn==23.0.0
python-dotenv==1.0.1
